+++ app/helpers/kitten/components_helper.rb
@@ -1,23 +1,22 @@
 module Kitten
   module ComponentsHelper
-    def render_group(group, title: nil)
-      title ||= default_title(group)
-      add_menu_group(group, title)
-      concat content_tag(:h2, class: 'k-Row karl-Title', id: group) { title }
-      render "kitten/groups/#{group}"
+    def render_category(category)
+      id = category[:id]
+      title = default_title(id)
+      add_menu_group(id, title)
+
+      content_tag(:h2, class: 'k-Row karl-Title', id: id) { title }
     end
 
     def render_component(component,
                          title: nil,
-                         description: nil,
-                         examples_display: :horizontal)
+                         description: nil)
       title ||= default_title(component)
       add_menu_component(component, title)
       render 'layouts/kitten/component',
              component: component,
              title: title,
-             description: description,
-             examples_display: examples_display
+             description: description
     end
 
     def example(title = nil)