+++ assets/javascripts/kitten/components/form/text-input-and-select.js
@@ -10,16 +10,23 @@ export const TextInputAndSelect = props => {
           selectOptions,
           selectName,
           selectValue,
+          onSelectChange,
+          errorOnSelect,
+          validOnSelect,
           appendSelect,
           ...textInputProps } = props
 
+  // TODO: use another prop that digits to handle
+  // `k-FormComposer__element--main` class.
   const inputWrapperClassName = classNames(
     'k-FormComposer__element',
     {
       'k-FormComposer__element--main': !props.digits
-    }
+    },
+    className,
   )
 
+  // TODO: handle key without currencyName prop
   const renderTextInputWrapper = () => {
     return (
       <div className={ inputWrapperClassName } key={ `${props.currencyName}1` }>
@@ -37,6 +44,9 @@ export const TextInputAndSelect = props => {
                          options={ selectOptions }
                          value={ selectValue }
                          tiny={ tiny }
+                         onInputChange={ onSelectChange }
+                         error={ errorOnSelect }
+                         valid={ validOnSelect }
                          disabled={ disabled } />
       </div>
     )
@@ -63,6 +73,10 @@ TextInputAndSelect.defaultProps = {
   selectOptions: [{ value: 'myValue', label: 'My label' }],
   selectName: null,
   selectValue: null,
+  onSelectChange: function() {},
+  errorOnSelect: false,
+  validOnSelect: false,
+  disabled: false,
   digits: null,
   appendSelect: false,
 }