+++ assets/javascripts/kitten/karl/form/radio-button.js
@@ -1,38 +1,29 @@
 import React from 'react'
 import { RadioButton } from 'kitten/components/form/radio-button'
 
-export const KarlRadioButtonError = () => {
-  return (
-    <div className="k-TextInputWithInfo">
-      <RadioButton error
-                   name="karlradiobuttonerror"
-                   text="Error default">
-      </RadioButton>
-      <RadioButton error
-                   name="karlradiobuttonerror"
-                   text="Error default checked"
-                   defaultChecked>
-      </RadioButton>
-      <p className="k-TextInputWithInfo__error">
-        Penatibus et magnis
-      </p>
-    </div>
-  )
-}
+export const KarlRadioButtonError = () => (
+  <div className="k-TextInputWithInfo">
+    <RadioButton error
+                 name="karlradiobuttonerror"
+                 text="Error default">
+    </RadioButton>
+    <RadioButton error
+                 name="karlradiobuttonerror"
+                 text="Error default checked"
+                 defaultChecked>
+    </RadioButton>
+  </div>
+)
 
-export const KarlRadioButtonContentError = () => {
-  return (
-    <div className="k-TextInputWithInfo">
-      <RadioButton error
-                   name="karlradiobuttoncontenterror"
-                   text="Error">
-         Quam multa enim, quae nostra causa numquam faceremus,
-         facimus causa amicorum! precari ab indigno, supplicare, tum
-         acerbius in aliquem invehi insectarique vehementius.
-      </RadioButton>
-      <p className="k-TextInputWithInfo__error">
-        Penatibus et magnis
-      </p>
-    </div>
-  )
-}
+export const KarlRadioButtonContentError = () => (
+  <div className="k-TextInputWithInfo">
+    <RadioButton error
+                 id="karl-radio-button-with-content-error"
+                 name="karl-radio-button-with-content"
+                 text="Error">
+       Quam multa enim, quae nostra causa numquam faceremus,
+       facimus causa amicorum! precari ab indigno, supplicare, tum
+       acerbius in aliquem invehi insectarique vehementius.
+    </RadioButton>
+  </div>
+)