+++ assets/javascripts/kitten/karl/organisms/loan-simulator.js
@@ -2,43 +2,43 @@ import defaultProps from 'kitten/hoc/default-props'
 import React from 'react'
 import LoanSimulator from 'kitten/components/simulators/loan-simulator'
 
-const nbsp = String.fromCharCode(160)
 const KarlLoanSimulator = defaultProps(LoanSimulator, {
-  amountLabel: 'J’ai besoin de',
-  amountPlaceholder: `Entre 200${nbsp}€ et 10${nbsp}000${nbsp}€`,
+  amountLabel: 'I need',
+  amountPlaceholder: 'Between $200 and $10,000',
   amountMin: 200,
   amountMax: 10000,
   amountName: 'amount',
-  initialAmount: null,
 
-  amountEmptyError: 'Veuillez renseigner un montant',
-  amountOutOfBoundsError: `Veuillez renseigner un montant compris entre
-                           200 et 10${nbsp}000${nbsp}€`,
+  amountOutOfBoundsError: 'Please choose an amount between $200 and $10,000',
 
-  installmentLabel: 'Je rembourse',
   installmentName: 'installment',
 
   durationName: 'duration',
 
-  sliderPlaceholder: 'Glisser pour sélectionner',
+  sliderPlaceholder: 'Slide to choose',
 
-  durationText: 'soit',
-  durationMin: 1,
-  durationMax: 36,
-  durationSymbol: 'mensualité',
-  durationSymbolPlural: 'mensualités',
-
-  currencySymbol: '€',
-  installmentSymbol: '€/mois',
-  locale: 'fr',
-
-  actionLabel: `C’est parti${nbsp}!`,
+  actionLabel: 'OK',
 })
 
 const KarlLoanSimulatorWithCommission = defaultProps(KarlLoanSimulator, {
   displayCommission: true,
-  commissionLabel: `Commission${nbsp}:`,
 
+  commissionRules: [
+    { durationMax: 9, rate: 0.03 },
+    { durationMax: 18, rate: 0.04 },
+    { durationMax: 24, rate: 0.05 },
+    { rate: 0.06 },
+  ],
+
+  initialAmount: 2500,
+  initialInstallment: 150,
+
+  actionLabel: null,
+})
+
+const KarlLoanSimulatorWithfeesExemption = defaultProps(KarlLoanSimulator, {
+  displayCommission: true,
+  feesExemption: true,
   commissionRules: [
     { durationMax: 9, rate: 0.03 },
     { durationMax: 18, rate: 0.04 },
@@ -59,4 +59,5 @@ const KarlLoanSimulatorWithError = defaultProps(KarlLoanSimulator, {
 
 export { KarlLoanSimulator,
          KarlLoanSimulatorWithCommission,
+         KarlLoanSimulatorWithfeesExemption,
          KarlLoanSimulatorWithError }