+++ CHANGELOG.md
@@ -7,13 +7,69 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: Add `BurgerIcon` and component.
 - Fix: Add `KisskissIcon` component.
 - Fix: Add `InformationIcon` component.
-- Feature: Add `Separator` component.
-- Fix: Install new grid on each components.
-- Fix: Delete example descriptions.
-- Fix: Redesign new subtitles and add border-bottom.
-- Fix: Install new grid on 10 columns.
-- Breaking change: Update button `min-width` class on `button-tiny`.
-- Feature: Add `margin` prop to `Title` component.
+
+## [11.0.0] - 2017-04-24
+
+Breaking changes:
+- Delete `SimulatorResult` component.
+- The `k-Select` mixin does not accept options anymore.
+  Please update `@import k-Select((…));` to `@import k-Select;`.
+- Remove `react-on-rails` module as dependency.
+  To be retro-compatible, add the dependency on your project's `package.json`:
+  ```
+  "dependencies": {
+    "react-on-rails": "~6.10.1"
+  }
+  ```
+
+Features:
+- Add `rounded="micro"` and `rounded="tiny"` to the `Card` component.
+- Add `ProjectCard` component.
+- Add `withoutPointerEvents` prop to `ButtonImage` component.
+- Add `CheckedCircleIcon`, `ClockCircleIcon`, `CrossCircleIcon`, `LockIcon` and
+  `WarningCircleIcon` components.
+- Add `tiny` prop to `TagList` component.
+- Update grid in `LegoGrid.Item` component.
+- Add `Container` component.
+
+Fixs:
+- Add `validOnSelect`, `disabled` props to the `TextInputAndSelect` component.
+- Add `valid`, `disabled` props to the `SelectWithState` component.
+- Add `checked-disabled` and `arrow-bottom-disabled` svgs.
+- Update disabled color on `TextInput`, `TextInputAndSelect`, `Button` and
+  `Select`.
+- Update `react_on_rails` gem dependency to `6.10.1`
+
+## [10.2.0] - 2017-04-11
+
+Feature:
+- Add fees exemption to `LoanSimulator` component.
+
+## [10.1.0] - 2017-04-11
+
+Fix:
+- Background `Alert` width.
+
+## [10.0.0] - 2017-04-10
+
+Breaking changes:
+- Use `k-Card--withVerticalPadding` and `k-Card--rounded` to have the same
+render of old `k-Card`.
+- Update button `min-width` class on `.k-Button--tiny`.
+
+Features:
+- Add modifiers on `k-Card` mixin.
+- Add `card` high-order component.
+- Add `normalLineHeight` prop to `Paragraph` component.
+- Add `margin` prop to `Title` component.
+- Add `onSelectChange` and `errorOnSelect` props to `TextInputAndSelect`
+component.
+- Add `Separator` component.
+
+Fixes:
+- Add missing checked disabled style for `RadioButton` component.
+- Set `line-height: normal` to `RadioButton/Checkbox/Switch` components.
+- Propagate `className` prop to `Grid` and `GridCol` components.
 
 ## [9.5.0] - 2017-03-30
 
@@ -39,12 +95,6 @@ Features:
 - Add `TagList` component.
 - Add `TagIcon` component.
 - Create `SupTitle` component.
-
-Fixes:
-- Replace images by empty images.
-- Delete description.
-- Redesign new subtitles and add border-bottom.
-- Install new grid on 10 columns.
 
 ## [9.2.0] - 2017-03-23